fix compatibility with Sphinx 1.0
Sphinx 1.0 (which we are using on our "new" RHEL6 Jenkins master)
doesn't handle multiple terms for a single glossary entry. The doc
really should have been written this way to begin with, rather than
duplicating the singular and plural version of the term in the glossary.

See also: http://stackoverflow.com/a/23853904/120202